Add files via upload
An example output file has been included in the repository to showcase the results generated by the Jupyter Notebook. This file contains UCSC Genome Browser links with custom tracks for visualizing guide RNA sequences.

Example Output File:

The output file includes the following columns:

Transcript_Name: The name of the transcript.
Guide_Sequence: The guide RNA sequence.
Gene_Name: The name of the gene.
Chromosome: Chromosome where the LncRNA is located.
Start: Start position of the LncRNA.
End: End position of the LncRNA.
Strand: The strand orientation of the guide RNA.
UCSC_URL: The UCSC Genome Browser link to visualize the guide RNA sequence.

This example demonstrates the format and structure of the output generated by running the notebook.